moving qupc functions to _skgraph.queries
