+++ openpnm/_skgraph/queries/__init__.py
@@ -1,1 +1,2 @@
 from ._funcs import *
+from ._qupc import *
